+++ src/App.tsx
@@ -8,10 +8,11 @@ import ErrorMessage from "./components/messages/ErrorMessage";
 import LoadingMessage from "./components/messages/LoadingMessage";
 import EmptyMessage from "./components/messages/EmptyMessage";
 import type { Movie } from "./types/movie";
+import { useMovies } from "./hooks/useMovies";
 
 function App() {
   const [themeName, setThemeName] = useState<ThemeName>("light");
-  const [movies, setMovies] = useState<Movie[]>([]);
+  const { movies, setMovies, refetch } = useMovies();
   const [newTitle, setNewTitle] = useState<string>("");
   const [newDirector, setNewDirector] = useState<string>("");
   const [newYear, setNewYear] = useState<number | "">("");
@@ -22,27 +23,6 @@ function App() {
   const [isLoading, setIsLoading] = useState<boolean>(false);
 
   const currentTheme = themes[themeName];
-
-  useEffect(() => {
-    const fetchMovies = async () => {
-      setIsLoading(true);
-      try {
-        const res = await fetch("/api/movies");
-        if (!res.ok) throw new Error("영화 데이터를 불러오지 못했습니다.");
-        const data: Movie[] = await res.json();
-        setMovies(data);
-      } catch (err: unknown) {
-        if (err instanceof Error) {
-          setError(err.message);
-        } else {
-          setError(String(err));
-        }
-      } finally {
-        setIsLoading(false);
-      }
-    };
-    fetchMovies();
-  }, []);
 
   const filteredMovies = useMemo(() => {
     return movies.filter((movie) =>
